Remove the -v option.
The verbose option relied upon a non-standard and undocumented implementation
of pthread_t. Unfortunately, I couldn't come up with a simple way to provide a
thread-local identifier, and I feel that the function of the verbose option was
too insignificant to warrant the added complexity in the code. It's only useful
for debugging purposes, and gdb can accomplish that better anyway.