Fix another error in README.
Shallot generates one modulus, and then searches for a value of e which
produces a hash value meeting the user-specified constraints.  This is not
the procedure which Tor uses, and public keys generated by Shallot can
easily be distinguished from public keys generated by Tor.  The reason
keypairs generated by Shallot are as secure as those generated by Tor
is that Shallot performs all of the sanity-checks which Tor does, and
possibly more.